change double to float

+++ moveit_ros/visualization/trajectory_rviz_plugin/include/moveit/trajectory_rviz_plugin/trajectory_display.h
@@ -68,7 +68,7 @@ public:
   void loadRobotModel();
 
   void load(const rviz_common::Config& config) override;
-  void update(float wall_dt, float ros_dt) override;
+  void update(double wall_dt, double ros_dt) override;
   void reset() override;
 
   // overrides from Display
